service/requirements.txt
flask
opencv-python-headless
numpy
pandas
tensorflow-addons
opencv-python
tqdm
matplotlib
pytest
tensorflow-datasets
